rename sub-package to `window`
