+++ susapad/controller/exception.py
@@ -1,5 +1,5 @@
 
-from susapad.windows import alert
+from susapad.window import alert
 
 
 def susapad_not_found(window):
