+++ susapad/window/settings/widgets/sensibility.py
